Fix based role #1

+++ src/controllers/department-controllers.js
@@ -1,8 +1,12 @@
 import Department from '../models/Department.js';
+import User from '../models/User.js';
 
 // Create department controller
 export const createDepartmentController = async (req, res) => {
     try {
+        const currentUser = await User.findById(req.user._id);
+        if (currentUser.isActived === false)
+            return res.status(403).json({ code: 403, message: 'Tài khoản tạm thời bị vô hiệu hóa' });
         const department = await Department.findOne({ departmentName: req.body.departmentName });
         if (department) return res.status(403).json({ code: 403, message: 'Tên phòng ban đã được sử dụng' });
         const newDepartment = new Department(req.body);
@@ -18,6 +22,9 @@ export const createDepartmentController = async (req, res) => {
 // Update department controller
 export const updateDepartmentController = async (req, res) => {
     try {
+        const currentUser = await User.findById(req.user._id);
+        if (currentUser.isActived === false)
+            return res.status(403).json({ code: 403, message: 'Tài khoản tạm thời bị vô hiệu hóa' });
         const department = await Department.findById(req.params.departmentId);
         if (!department) return res.status(404).json({ code: 404, message: 'Không tìm thấy phòng ban' });
 
@@ -40,6 +47,9 @@ export const updateDepartmentController = async (req, res) => {
 // Activate department controller
 export const activateDepartmentController = async (req, res) => {
     try {
+        const currentUser = await User.findById(req.user._id);
+        if (currentUser.isActived === false)
+            return res.status(403).json({ code: 403, message: 'Tài khoản tạm thời bị vô hiệu hóa' });
         const status = req.body.status;
         await Department.findByIdAndUpdate(req.params.departmentId, { $set: { status: status } });
         if (status === true) {
@@ -56,6 +66,9 @@ export const activateDepartmentController = async (req, res) => {
 // Delete department controller
 export const deleteDepartmentController = async (req, res) => {
     try {
+        const currentUser = await User.findById(req.user._id);
+        if (currentUser.isActived === false)
+            return res.status(403).json({ code: 403, message: 'Tài khoản tạm thời bị vô hiệu hóa' });
         const department = await Department.findById(req.params.departmentId);
         if (!department) return res.status(404).json({ code: 404, message: 'Không tìm thấy phòng ban' });
 
@@ -70,6 +83,9 @@ export const deleteDepartmentController = async (req, res) => {
 // Delete many departments permanently controller
 export const deleteManyDepartmentController = async (req, res) => {
     try {
+        const currentUser = await User.findById(req.user._id);
+        if (currentUser.isActived === false)
+            return res.status(403).json({ code: 403, message: 'Tài khoản tạm thời bị vô hiệu hóa' });
         const arrayId = req.body.arrayId;
         await Department.deleteMany({ _id: arrayId });
         res.status(200).json({
@@ -85,6 +101,9 @@ export const deleteManyDepartmentController = async (req, res) => {
 // Get all list department and pagination controller
 export const getAllDepartmentController = async (req, res) => {
     try {
+        const currentUser = await User.findById(req.user._id);
+        if (currentUser.isActived === false)
+            return res.status(403).json({ code: 403, message: 'Tài khoản tạm thời bị vô hiệu hóa' });
         let { page, limit, search } = req.query;
         if (!page) page = 1;
         if (!limit) limit = 5;
@@ -115,6 +134,9 @@ export const getAllDepartmentController = async (req, res) => {
 // Get department by ID controller
 export const getDepartmentByIdController = async (req, res) => {
     try {
+        const currentUser = await User.findById(req.user._id);
+        if (currentUser.isActived === false)
+            return res.status(403).json({ code: 403, message: 'Tài khoản tạm thời bị vô hiệu hóa' });
         const department = await Department.findById(req.params.departmentId);
         if (!department) return res.status(404).json({ code: 404, message: 'Không tìm thấy phòng ban' });
         res.status(200).json({ code: 200, data: department });
